+++ decidim-surveys/app/forms/decidim/surveys/admin/survey_question_form.rb
@@ -10,24 +10,19 @@ module Decidim
         attribute :position, Integer
         attribute :mandatory, Boolean, default: false
         attribute :question_type, String
-        attribute :options, Array[SurveyQuestionAnswerOptionForm]
+        attribute :answer_options, Array[SurveyQuestionAnswerOptionForm]
         attribute :deleted, Boolean, default: false
 
         translatable_attribute :body, String
+        translatable_attribute :description, String
 
         validates :position, numericality: { greater_than_or_equal_to: 0 }
         validates :question_type, inclusion: { in: SurveyQuestion::TYPES }
         validates :body, translatable_presence: true, unless: :deleted
 
         def map_model(model)
-          self.options = model.answer_options.each_with_index.map do |option, id|
-            [id + 1, option]
-          end
-        end
-
-        def options=(value)
-          @options = value.map do |id, option|
-            SurveyQuestionAnswerOptionForm.new(option.merge(id: id.to_s.to_i))
+          self.answer_options = model.answer_options.each_with_index.map do |option, id|
+            SurveyQuestionAnswerOptionForm.new(option.merge(id: id + 1))
           end
         end